refactor:仪表板编辑时，文字组件、矩形组件、过滤组件样式设置面板强制在画布区域内
